# docs/swagger.yaml
basePath: /api/v1
definitions:
  apperror.AppError:
    properties:
      code:
        $ref: '#/definitions/apperror.Code'
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
  apperror.Code:
    enum:
    - NOT_FOUND
    - CONFLICT
    - VALIDATION_ERROR
    - UNAUTHORIZED
    - FORBIDDEN
    - INTERNAL_ERROR
    - BAD_REQUEST
    - UNPROCESSABLE
    type: string
    x-enum-varnames:
    - CodeNotFound
    - CodeConflict
    - CodeValidation
    - CodeUnauthorized
    - CodeForbidden
    - CodeInternal
    - CodeBadRequest
    - CodeUnprocessable
...
host: localhost:8080
info:
  contact:
    email: support@edossier.com
    name: API Support
  description: e-Dossier School Management System API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: e-Dossier API
  version: "1.0"
paths:
...
  /reports/dashboard:
    get:
      description: Retrieve dashboard statistics for the authenticated user's state/school
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperror.AppError'
      summary: Get dashboard stats
      tags:
      - reports
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" + space + JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"